Use python3.9 for testing cm-script

+++ cm-mlops/script/get-imagenet-val/run.sh
@@ -8,10 +8,12 @@ if [ ${CM_IMAGENET_FULL:-no} == "yes" ]; then
   echo "CM_DATASET_PATH=${IMAGENET_PATH}" > tmp-run-env.out
   exit 0
 fi
+test $? -eq 0 || exit 1
 
 if [ -f "ILSVRC2012_img_val_500.tar" ]; then
   rm -f ILSVRC2012_img_val_500.tar
 fi
+test $? -eq 0 || exit 1
 
 wget -nc https://www.dropbox.com/s/57s11df6pts3z69/ILSVRC2012_img_val_500.tar --no-check-certificate
 test $? -eq 0 || exit 1
